requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
cssselect==1.1.0
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.43.0
django-crispy-forms==1.10.0
django-markdown-deux==1.0.5
django-tinymce4-lite==1.8.0
idna==2.10
isort==5.6.4
jsmin==2.2.2
lazy-object-proxy==1.5.1
lxml==4.6.1
markdown2==2.3.10
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
pyquery==1.4.3
python3-openid==3.2.0
pytz==2020.4
readtime==1.1.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
